requirements.txt
moviepy==1.0.3
edge-tts
openai-whisper
requests
python-dotenv
pillow